test: move null order_by regression tests to base test classes
Move multi-null order_by tests from PyArrow-specific test files to
FrameAggregateTestBase and WindowAggregationTestBase so all compute
frameworks (PyArrow, Pandas, DuckDB, Polars, SQLite) inherit them.
